# server/openapi.yaml
openapi: 3.1.0
info:
  title: WealthGenie API
  version: 1.0.0
  description: >
    Production-grade financial planning and investment advice engine API.
    Provides automated portfolio recommendation, Monte Carlo wealth projection,
    Indian tax optimization (FY2025-26), and XIRR goal calculations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local Development Server
  - url: https://api.wealthgenie.io
    description: Production Cluster Gateway
paths:
...
  /api/profile/build:
    post:
      operationId: createProfile
      summary: Create / Build Profile
      description: Creates or updates the user's financial profile.
      tags:
        - User Profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
      responses:
        '200':
          description: Profile saved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/profile/{id}:
    put:
      operationId: updateProfile
      summary: Update Profile by ID
      description: Updates an existing profile by document ID with ownership check.
      tags:
        - User Profile
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
      responses:
        '200':
          description: Profile updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/tax/compare:
    post:
      operationId: compareTaxRegimes
      summary: Compare Tax Regimes
      description: Compares Old vs New tax regime and recommends optimal regime.
      tags:
        - Tax Engine
      responses:
        '200':
          description: Tax regime comparison complete
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request / validation error
      content:
        application/json:
          example:
            error: Invalid request parameters
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          example:
            error: Access denied. No token provided.
    Forbidden:
      description: Access forbidden / ownership check failed
      content:
        application/json:
          example:
            error: Access denied. Resource ownership mismatch.
...
  schemas:
    ProfileRequest:
      type: object
      required: [age, monthly_income, monthly_savings, risk_tolerance, total_ctc, basic_component, monthly_take_home]
      properties:
        age:
          type: integer
          minimum: 18
          maximum: 80
          example: 30
        monthly_income:
          type: number
          minimum: 1000
          maximum: 100000000
          example: 150000
        monthly_savings:
          type: number
          minimum: 500
          example: 50000
        total_ctc:
          type: number
          minimum: 100000
          maximum: 1000000000
          example: 2400000
        basic_component:
          type: number
          minimum: 20000
          maximum: 1000000000
          example: 1200000
        monthly_take_home:
          type: number
          minimum: 1000
          example: 140000
        sold_property_amount:
          type: number
          minimum: 0
          maximum: 10000000000
          example: 2000000
        has_lump_sum:
          type: boolean
          default: false
          example: true
        lump_sum_amount:
          type: number
          minimum: 0
          maximum: 10000000000
          example: 2000000
        risk_tolerance:
          type: string
          enum: [Conservative, Moderate, Aggressive]
          example: Moderate
...