# server/routes/profile.js
import { Router } from 'express';
import { verifyJWT } from '../middleware/authMiddleware.js';
import { asyncHandler, createError } from '../middleware/errorHandler.js';
import { validate, profileSchema } from '../validation/schemas.js';
import { computeTax, getTaxSlab, compareTaxRegimes } from '../services/taxEngine.js';
import { getRiskProfile } from '../services/riskProfiler.js';
import FinancialProfile from '../models/FinancialProfile.js';
import { delCache, redisClient, redisAvailable } from '../config/redis.js';
import { idempotency } from '../middleware/idempotency.js';

const router = Router();

// Profile creation throttle — max profiles per user per hour
// STATELESS: Uses shared Redis counter. No per-process in-memory state.
const PROFILE_RATE_LIMIT = 10;

async function checkProfileRateLimit(userId) {
  if (!redisAvailable || !redisClient) {
    // Without Redis, skip throttle rather than diverging per-process state.
    return true;
  }
  try {
    const key = `profile:ratelimit:${userId}`;
    const count = await redisClient.incr(key);
    if (count === 1) await redisClient.expire(key, 3600);
    return count <= PROFILE_RATE_LIMIT;
  } catch {
    // Redis failure — allow request through to avoid blocking users.
    return true;
  }
}

/**
 * POST /api/profile/build [Protected]
 * Builds a financial profile with tax computation and risk profiling.
 */
router.post('/build', verifyJWT, idempotency(), validate(profileSchema), asyncHandler(async (req, res) => {
  // Throttle: prevent unbounded profile creation
  if (process.env.DISABLE_RATE_LIMIT !== 'true' && !checkProfileRateLimit(req.user.userId)) {
    throw createError(429, `Profile rate limit for user ${req.user.userId}`,
      `Too many profile submissions. Maximum ${PROFILE_RATE_LIMIT} per hour.`);
  }

  const {
    monthly_income, age, monthly_savings, regime, investment_horizon,
    liquid_savings, existing_debt, dependents, emergency_fund_months,
    risk_tolerance, goal_type,
    total_ctc, basic_component, monthly_take_home, sold_property_amount,
    has_lump_sum, lump_sum_amount
  } = req.body;

  const annualIncome = monthly_income * 12;
  const taxRegime = regime || 'new';

  const safeTotalCTC = Number(total_ctc) || annualIncome;
  const safeBasicComponent = Number(basic_component) || (safeTotalCTC * 0.5);
  const safeMonthlyTakeHome = Number(monthly_take_home) || monthly_income;
  const safeSoldPropertyAmount = Number(sold_property_amount) || 0;
  const safeHasLumpSum = Boolean(has_lump_sum);
  // Defensive zeroing: force lump sum to 0 if has_lump_sum flag is false
  const safeLumpSumAmount = safeHasLumpSum ? (Number(lump_sum_amount) || 0) : 0;

  // Compute tax with basicSalary passed to calculate 80CCD(2) employer-NPS deduction cap
  const taxDeductions = { basicSalary: safeBasicComponent };
  const taxResult = computeTax(annualIncome, taxRegime, taxDeductions);
  const marginalRate = getTaxSlab(annualIncome, taxRegime, taxDeductions);
  const taxComparison = compareTaxRegimes(annualIncome, taxDeductions);

  // Compute risk profile incorporating available lump sum & property sale liquidity
  const riskProfile = getRiskProfile(
    age, annualIncome, investment_horizon, 0, dependents || 0,
    monthly_savings, existing_debt || 0, safeLumpSumAmount, safeSoldPropertyAmount
  );

  // Investable amount separation: monthly vs one-time lump sum
  const investableAmount = monthly_savings;

  // Self-check invariant: riskScore must align with riskCategory
  const SCORE_RANGES = {
    'Aggressive': [80, 100], 'Moderate-Aggressive': [60, 79],
    'Moderate': [40, 59], 'Conservative-Moderate': [20, 39], 'Conservative': [0, 19],
  };
  const expectedRange = SCORE_RANGES[riskProfile.category];
  if (expectedRange && (riskProfile.riskScore < expectedRange[0] || riskProfile.riskScore > expectedRange[1])) {
    console.error(
      `[Profile INVARIANT VIOLATION] riskScore ${riskProfile.riskScore} does not match `
      + `category '${riskProfile.category}' (expected ${expectedRange[0]}-${expectedRange[1]}).`
    );
  }

  // Save to MongoDB
  const profile = await FinancialProfile.create({
    userId: req.user.userId,
    income: monthly_income,
    age,
    savings: monthly_savings,
    annualIncome,
    taxSlab: marginalRate,
    effectiveTaxRate: taxResult.effectiveRate,
    taxRegime,
    riskCategory: riskProfile.category,
    riskScore: riskProfile.riskScore,
    riskDescription: riskProfile.description,
    recommendedEquityAllocation: riskProfile.recommendedEquityAllocation,
    investableAmount,
    investmentHorizon: investment_horizon,
    liquid_savings: liquid_savings || 0,
    existing_debt: existing_debt || 0,
    dependents: dependents || 0,
    emergency_fund_months: emergency_fund_months || 0,
    risk_tolerance: risk_tolerance || 'Moderate',
    goal_type: goal_type || 'wealth-building',
    totalCTC: safeTotalCTC,
    basicComponent: safeBasicComponent,
    monthlyTakeHome: safeMonthlyTakeHome,
    soldPropertyAmount: safeSoldPropertyAmount,
    hasLumpSum: safeHasLumpSum,
    lumpSumAmount: safeLumpSumAmount,
  });

  // Invalidate ALL chatbot system prompt caches for this user
  try {
    const prefix = `chat:sysprompt_v3:${req.user.userId}:`;
    await delCache(prefix + profile._id);
    const prevProfile = await FinancialProfile.findOne({
      userId: req.user.userId,
      _id: { $ne: profile._id },
    }).sort({ createdAt: -1 }).lean();
    if (prevProfile) await delCache(prefix + prevProfile._id);
  } catch (redisErr) {
    console.warn('[Profile] Cache invalidation failed (non-critical):', redisErr.message);
  }

  res.status(201).json({
    profileId: profile._id,
    taxSlab: marginalRate,
    effectiveTaxRate: taxResult.effectiveRate,
    taxDetails: taxResult,
    taxComparison,
    riskCategory: riskProfile.category,
    riskScore: riskProfile.riskScore,
    riskDescription: riskProfile.description,
    recommendedEquityAllocation: riskProfile.recommendedEquityAllocation,
    annual_income: annualIncome,
    investable_amount: investableAmount,
    investable_amount_monthly: monthly_savings,
    investable_amount_onetime: safeLumpSumAmount,
    total_ctc: safeTotalCTC,
    basic_component: safeBasicComponent,
    monthly_take_home: safeMonthlyTakeHome,
    sold_property_amount: safeSoldPropertyAmount,
    has_lump_sum: safeHasLumpSum,
    lump_sum_amount: safeLumpSumAmount,
  });
}));

/**
 * PUT /api/profile/:profileId [Protected]
 * Updates a financial profile with optimistic concurrency control (OCC).
 */
router.put('/:profileId', verifyJWT, validate(profileSchema), asyncHandler(async (req, res) => {
  const { profileId } = req.params;
  const expectedVersion = req.body.version;

  const profile = await FinancialProfile.findOne({ _id: profileId, userId: req.user.userId });
  if (!profile) {
    throw createError(404, 'Profile not found or access denied', 'Profile not found.');
  }

  // 1. Fail-fast version check
  if (expectedVersion !== undefined && profile.__v !== expectedVersion) {
    return res.status(409).json({
      error: 'Conflict',
      message: 'Version conflict. Document has been modified concurrently by another process.'
    });
  }

  const {
    monthly_income, age, monthly_savings, regime, investment_horizon,
    liquid_savings, existing_debt, dependents, emergency_fund_months,
    risk_tolerance, goal_type,
    total_ctc, basic_component, monthly_take_home, sold_property_amount,
    has_lump_sum, lump_sum_amount
  } = req.body;

  const safeTotalCTC = Number(total_ctc) || (monthly_income * 12);
  const safeBasicComponent = Number(basic_component) || (safeTotalCTC * 0.5);
  const safeMonthlyTakeHome = Number(monthly_take_home) || monthly_income;
  const safeSoldPropertyAmount = Number(sold_property_amount) || 0;
  const safeHasLumpSum = Boolean(has_lump_sum);
  const safeLumpSumAmount = safeHasLumpSum ? (Number(lump_sum_amount) || 0) : 0;

  // Apply edits
  profile.income = monthly_income;
  profile.age = age;
  profile.savings = monthly_savings;
  profile.annualIncome = monthly_income * 12;
  profile.taxRegime = regime || 'new';
  profile.investmentHorizon = investment_horizon;
  profile.liquid_savings = liquid_savings || 0;
  profile.existing_debt = existing_debt || 0;
  profile.dependents = dependents || 0;
  profile.emergency_fund_months = emergency_fund_months || 0;
  profile.risk_tolerance = risk_tolerance || 'Moderate';
  profile.goal_type = goal_type || 'wealth-building';
  profile.totalCTC = safeTotalCTC;
  profile.basicComponent = safeBasicComponent;
  profile.monthlyTakeHome = safeMonthlyTakeHome;
  profile.soldPropertyAmount = safeSoldPropertyAmount;
  profile.hasLumpSum = safeHasLumpSum;
  profile.lumpSumAmount = safeLumpSumAmount;

  // Recompute tax & risk profile
  const taxDeductions = { basicSalary: safeBasicComponent };
  const taxResult = computeTax(profile.annualIncome, profile.taxRegime, taxDeductions);
  const marginalRate = getTaxSlab(profile.annualIncome, profile.taxRegime, taxDeductions);
  const riskProfile = getRiskProfile(
    age, profile.annualIncome, investment_horizon, 0, profile.dependents || 0,
    monthly_savings, profile.existing_debt || 0, safeLumpSumAmount, safeSoldPropertyAmount
  );

  profile.taxSlab = marginalRate;
  profile.effectiveTaxRate = taxResult.effectiveRate;
  profile.riskCategory = riskProfile.category;
  profile.riskScore = riskProfile.riskScore;
  profile.riskDescription = riskProfile.description;
  profile.recommendedEquityAllocation = riskProfile.recommendedEquityAllocation;
  profile.investableAmount = monthly_savings;

  try {
    await profile.save();
  } catch (err) {
    if (err.name === 'VersionError') {
      return res.status(409).json({
        error: 'Conflict',
        message: 'Version conflict. Document has been modified concurrently.'
      });
    }
    throw err;
  }

  res.json({
    status: 'success',
    profile
  });
}));

export default router;